ledslie/requirements.txt
flask
flask_mqtt
pillow
msgpack-python
serial
